+++ src/Annotation/Preferred.php
@@ -13,6 +13,7 @@ namespace Liip\MetadataParser\Annotation;
  * @Annotation
  * @Target({"METHOD", "PROPERTY"})
  */
+#[\Attribute(\Attribute::TARGET_METHOD | \Attribute::TARGET_PROPERTY)]
 final class Preferred
 {
 }